api/requirements.txt
faiss-cpu
fastapi-cli
fastapi
langchain
langchain_community
langchain_openai
langgraph
pypdf
fpdf
openai
markdown
unstructured
azure-search-documents==11.6.0b3
uvicorn
typing_extensions
python-dotenv
pdfminer
pydantic
python-pptx
python-multipart
opentelemetry-api
diagrams
mermaid-python
azure-storage-blob
azure-core
pillow